Trust the OS.
I spent a bunch of time yesterday debugging this with Emanuele on
multitrack-app1. Some issues were because I didn't understand MQs, some
were because of resource limits, and some were because of bugs.

This removes the overly-complex bit to handle the ring buffer. I added
it because I didn't really trust the OS to do its job, which is clearly
dumb.