Fixed two bugs, one of which caused all our problems.
The NGX_LINEFEED_CHANGE here is because len is unsigned, causing the
comparison against 0 to always be false. I'm not sure if this would ever
actually come up, but this version expresses the intention.

The second change increases the buffer size for log messages to
correctly accomodate for high concentrations of unprintable characters.